Make sure to clear out users when logging out
Previously, the current user was being cached, so if a user logged out
and a different user logged in, they'd appear to be the first user until
they refreshed the browser. It doesn't look like it's good enough to
just call unloadRecord on the currentUser model (I think it has to do
with the fact that we're tricking Ember Data into thinking that the
current user has an id of `current`).

Let's just unload all users when we logout. I don't think this is so
terrible. We should probably be unloading everything on logout anyway.

[#31]